Add healthchecking route for ECS autoscaling
AWS target groups use the healthchecking route to determine whether
the app is ready for accepting incoming requests or not.  Once
the app is ready, it will be registered with the load balancer to
start receiving requests.

At the moment this healthcheck is very simple, as our app doesn't
have any complex behaviour.