Use ES6-style imports to get gov-frontend working
This commit contains the first webpacker-style imports for assets - so far,
only javascript. This runs the recommended 'init' command for the
gov-frontend npm package, and doesn't error. It's difficult to see if it's
truly 'working' before the styling assets are included as well.

See guide at https://github.com/alphagov/govuk-frontend/blob/master/
docs/installation/installing-with-npm.md#importing-javascript , which explains
this other option.