Allow Jenkins to run migrations
We call a scheduled task which will do the migrations for us in Jenkins.
Since we are running the migrations before making the task live, we will
need to keep the migrations non destructive.

We decided that this is ok, since we do this anyway for a blue green
setup.

We will look into making the new task live as soon as we have pushed a
new image into the ECR repository later on.