Verify email is in gov domain (#23)
This commit uses a more conventional ActiveModel validation, which 
gets us the error messages, fails the save, etc. for 'free'.

There's some hidden behaviour here, but it shouldn't be surprised to
people familiar with Rails, and doesn't require Devise knowledge to
parse either. There's a validation method on User, and corresponding
tests to explain the behaviour.

This approach has the added benefit of extracting the whitelisting
check out to a use case, which means it's easier to test, change in
the future, etc, etc.

I don't think 'administrator' is a good name for the particular actor
we're describing here - someone who's actually running GovWifi itself,
and wants to control who can sign up - but at the time of this commit,
we've not yet come up with something better.