Add use case to view RADIUS IPs from env-vars
This commit adds a bare-bones use case, deliberately not using a
gateway to keep things simple. Given we've talked about which IPs
we'll show to users, you could argue one should be added.

This implementation is pretty tied to there only being two RADIUS
locations (and indeed, to the presentation of IPs to organisations
being the strategy of "just show 'em all of them").

I don't foresee that changing in the short-to-medium term, but could
change long-term.