Tabify logged-in landing, add placeholder content
This commit introduces three tabs - home, locations and team - that
will hopefully help to make it clearer to a user what's happening once
they've logged in.

I think this version of the content is too dense to guide a user, and
I'm not happy with a lot of the language, we can refine it from here.