Improve UX of add IP page
This commit changes various markup and messages to better illustrate
to users the purpose and functionality of the now "Add Locations" page.

It also modifies the feature specs that were checking against these
pages.

Notably this changes the language from "adding IPs" to "adding
locations", which I'm not sure if correct. Feels like we have
divergent language around this concept.