Only validate email on create (#132)
We want to make sure users have an email that's from .gov (or other
whitelisted) domains, so that not anyone can sign up for GovWifi.

In the 'old' GovWifi system - the one where the user's interface is a 
bunch of emails - users might hand over control to an external
provider (e.g. BT), but those organisations aren't on the whitelist
(and can't be in some cases - we can't let everyone from BT sign up!)

Given we're bringing in data from that 'old' system, there's emails 
that aren't covered by the whitelist already, so this is necessary for 
migrating the data across.