Correctly show/hide MOU prompt, add specs
This commit implements the actual logic for showing the prompt -
whether the logged in user has already uploaded an MOU or not - and
removes the fake logic and artifacts that were previously in place.

It also folds some of the planned specs into the existing MOU feature
specs, to take advantage of their setup steps (and because it makes
sense to have them here). Could argue that there shouldn't be a
separate spec for the prompting at all.