Run aws update-service command to deploy
From the AWS docs:
"If your updated Docker image uses the same tag as what is in the existing
task definition for your service (for example, my_image:latest), you do
not need to create a new revision of your task definition.

You can update the service using the procedure below, keep the current
settings for your service, and select Force new deployment.

The new tasks launched by the deployment pull the current image/tag
combination from your repository when they start."

Pretty sure this is the model we're following - tagging a docker image as
'latest' or similar and then wanting to deploy it.

Notably this command run locally leaves an extra container running - this
commit might need an autoscaling-down for staging to accompany it.