Fix layout of main navigation to be consistent
The admin main navigation needs to match the product page and the tech
docs.  We are unable to use exactly the same styles as the tech docs and
product page have different stylesheets baked in.

This overwrites some of the default styles for the gov-uk frontend to
match these other 2 apps.