Separate out new multi-IP page
This commit comments out the javascript, points the 'add an IP to
this location' link back to the original page, and fixes the
corresponding tests.

This is to support splitting out the 'add an ip' flows into new
and existing location, which should make adding on the 'add more ips'
functionality smoother.