Search logs by location (#404)
This commit adds a third filter on log searching - location - that is 
a little more complex than the other two, as it requires 
interrogating Locations (an ActiveRecord concept) rather than just 
filtering on a column in Sessons.

To this end, some of the logic remains in the controller, but after a 
few cycles of refactoring, it should be pretty clear what's going on. 

This commit also introduces the views, specs and other supporting 
parts of this feature.