Remove "recent activity" stat  (#470)
The activity stat is causing the app to respond slowly, as the 
database query underneath can take 10+ seconds if the given 
organisation has many (20+ IPs). In the long-term, optimising this 
query is desirable, but in the short-term, it's not acceptable for 
the app to respond slowly.

This commit, in part, reverts PRs #409 and #448, but isn't a 
straight revert. It preserves some unrelated changes (spec 
improvements, folder name changes, styling).