Redirect the user to the support form
If a user's organisation has been removed and they no longer belong to any
organisation we should forward them to a support form. We don't expect to
see this happen as it would mean an admin has deleted an entire org.
Presents the user with a flash message explaining how to report the issue.