Only link to logged-in support when logged in
This re-works https://github.com/alphagov/govwifi-admin/pull/824

I think the distinction between support and admin support is a bit
confusing. Especially when admin support already links to end user
support.

We can fix the problem about non-signed-in users getting to the feedback
from by only linking to it if a user is signed in instead.

The IA for a signed-in user looks a bit like this:
```
            Support
                |
    —————————————————————————
    |           |           |
End user   Admin user   Raise ticket
```

For a non-signed-in user it looks like this:
```
```
       Support
          |
    —————————————
    |           |
End user   Admin user
                |
             Sign in
                |
          Raise a ticket