Make locations table three columns
When each column of the table maps to a column in the database you end
up showing states which don’t really make sense, because the alternative
would be impossible.

For example:
> Not available until 6am | No traffic received

Here the second statement is redundant, because it’s impossible for the
IP address to receive traffic until it is active.

This results in the UI repeating a lot of text, without efficiently
conveying a lot of information.

This commit simplifies the table into three columns:
> Thing | Status of thing | Remove thing

By being more dynamic in describing the status of each thing we can
be a bit more judicious about which information the users needs to see,
thereby increasing the information density.

And it looks a lot tidier.