Put a grey border above the first row of the table
The first row of the table blends with the controls above a bit.

We don’t want to add a lot of white space here because it will weaken
the visual relationship between the location and its IP addresses.

So a border adds a bit more structure, without losing the relationship.