Only show search once the user has multiple locations
The search box filters the items on the page. You’ll never have an item
off the bottom of the screen until another item comes before it. So you
need a minimum of 2 items before the on-page search is at all useful.

For those who don’t, or might never have more than one location it’s
just another distraction which makes the interface less straightforward.