Always link to add location from top right
If you want to add IP addresses to an existing location you can do that
from the location itself.

This means that each button does one, clear thing, rather than having
multiple button which do the same, or only slightly different thing.

It also means that adding a new location should be easier to find, at
the moment you have to know that it’s an ‘alternative’ link at the
bottom of the ‘Add IP address’ page.