Separate adding locations from adding IP addresses
There are two problems with the journey at the moment:
- there are multiple places to get to the add IP address page from, and
  it’s not clear which one you should use (this adds cognitive load)
- adding a new location is hidden behind the button that says ‘Add IP
  address’
- there’s a `<select>` control for choosing a location, which we
  generally try to avoid

This commit simplifies the pages by having one button to add a location,
and a button for each location that lets you add IP address

Making this journey simpler makes a bit of room to add some contextual
help and hint text that should help people complete these tasks.

I’ve also made the ability to add 5 IP addresses available (not just
when you’re adding a location) because it felt useful.