Remove stuff that’s already in the documentation
There’s a lot of confusing duplication between this page and the
documentation, which:
- makes it harder to pick out the relevant information on this page
- makes it harder for the user to know where they should be looking
  for information

We think that the documentation can’t and shouldn’t be eliminated,
because it covers stuff that’s not included on these pages. Users will
have to have both open side by side, so this page should focus on the
stuff that we can only show to someone’s who’s logged in.

I’ve had a chat with the technical writer (Pippa) about this and she
agrees.

This commit also removes the info about IP addresses and documentation,
because that duplicates information that’s already on the location page.