test: workaround an issue with libsassc and min/max() in stylesheets
In a nutshell the new libsassc compiler doesn't support the
min()/max() syntax because they'd been removed off the CSS spec, taken
over by SASS as SASS-specific functions, and now they have been
re-introduced in the spec and so confusion arise and the test suite crashes.

The workaround is to disable the compiler in test environments to
avoid RSpec and libsassc somehow choking each other.

See the full detective workfor more details:

https://github.com/alphagov/govuk-frontend/issues/1350