Cap `rails-sass` dependency to less than v6.0.0
`sassc` has problems building on Concourse[1], so we’ve stuck with
the older `rails-sass`.

However, as of version 6.0.0 `rails-sass` have moved to using `sassc`
under the hood[2], reintroducing the problem.

This commit limits the version of `rails-sass` installed to the newest
version that still uses the older, Ruby-based dependencies for SASS
processing.

1. https://github.com/alphagov/govwifi-admin/commit/85586d21a2370cbc3d0601a611f10d0e7540fe79
2. https://github.com/rails/sass-rails/releases/tag/v6.0.0