Fix error summary shown before entering IPs bug
The error summary should only show if there are IP addresses which
couldn’t be persisted. It’s supposed to do this by filtering out blank
ones and seeing if there are any left. It was doing the opposite,
leaving only the blank ones. In other word the presence of any blank
IP addresses (like the 5 we create when you view the page) was causing
the error to show.

However this confuses the page if the user hasn’t submitted anything,
because the model doesn’t know whether there’s a `GET` or a `POST` going
on. So I’ve had to put a bit of logic in the view as well to check
which controller method is rendering the page.