Add an authenticated URL to download the poster
We need to link to the poster from a public place, but we still only
want people who are signed in to be able to download it. So we can’t
link to the asset directly.

What we can do instead is add an endpoint which serves the asset but
has authentication.