feature: re-introduce is_super_admin on the users table
We moved away from this attribute in 6cd35 which removed the attribute
and used a super-organisation instead. Since this assimilation (i.e
super-admins are part of a super-org) is misleading and impractical,
restore the attribute in preparation of migrating superadmin
functionalities into a separate UI, rather than patching onto the
regular UI for (normal) organisations using GovWifi.