fix: ensure a classic super-admin can use the normal interface
The logic in master uses a `super_admin?` helper which is bound to the
`current_organisation` rather than the user. This was overlooked in
the new code and therefore wrongly redirected to the classic
superadmin views as soon as the user was a super-admin, even if they
were trying to view a normal organisation they belong to.