fix: correctly point the label to the orgs auto-complete
Setting the label's `for` attribute to `:name` eventually translates
to `user_organisations_attributes_0_name` which we don't want to
enforce on the auto-complete as we use it in other places and forms.

Instead use the ID used by `accessible-autoselect` to find and
override the select control as this makes the label point correctly to
both versions (i.e normal and enhanced).

Also make sure it uses the correct font.