Change the `password_present` method to check for nil, rather than relying on presence.
This is because we only added this check to ensure we are able to make updates to the User
model for changes which don't include password. However, checking for presence of password
also makes it so that empty passwords are not considered valid since they won't be validated.

Changing this method to check for nil instead is more correct – empty strings are still validated
as they aren't nil, and updating the User model without passing password will skip password
validation as per the original intent of this method.