fix: simplify the reset 2FA spec
We can avoid a lot of mocking and messing about by just setting the
`otp_secret_key` to a value, as demonstrated by the users factory. Use
this instead, which means no more mocking or manually going once or
twice through the signup/setup-2fa process depending on the test case.