fix: add more tests around the 2FA edits
And introduce some controller tests as they're a lot more
accurate (and marginally faster) to verify
permissions/redirect/rendering logic.

Also try and simplify the controller by moving the logic of "are you
allowed to manage this person's account" inside the User class where
it lives closer to the required attributes and add some tests for it.