BAU: Run auto rebase for dependabots at night
The admin repo has a fair bit of movement and everytime we merge a PR, all pending dependabot PRs will automatically rebase and trigger a PR build on concourse. This can significantly impact build run time when you are regularly pushing on a branch.
So, for the time being, we are only running rebases at 3AM, which should also decrease the number of rebases and jobs triggered.