bau: replace govuk-lint
Following the instructions on the deprecated govuk-lint gem[1], remove
govuk-lint in favour of `rubocop-govuk` and `scss-lint-govuk`. The
main changes are:

- double-quotes everywhere: strings are frozen by default in upcoming
versions of Ruby and we have the frozen_string_literals magic comment
to enable it anyway;
- using Time.zone.today instead of Date.today: timezone-aware;
- Rails.root.joins is better than manually joining;
- all previous migrations have been ignored as it's so much work to
backport them into a Rubocop-friendly form.

Also some minor versions updates thrown in by `bundle update`.

[1]: https://github.com/alphagov/govuk-lint#deprecated