BAU - allow manual trigger of PR jobs
Occasionally a concourse PR build job will fail for reasons unrelated with the PR itself (network timeout, for example). This means we usually need to do a workaround of force pushing an "empty" commit to re-trigger the job.
This PR allows for concourse jobs to be triggered via the concourse GUI or via fly by pinning the version