Add the screen for selecting a 2FA method.
The 2FA methods include 'email' and 'app', the first of which is new and
will be implemented in a subsequent commit. This screen doesn't actually
do anything - it just allows you to select a 2FA method - business logic
will come later.

We now also redirect to this new screen, if 2FA hasn't been set up yet,
instead of the previous app 2FA setup screen.

Co-authored-by: Alan Carter <alan.carter@digital.cabinet-office.gov.uk>