fix: ensure we turn off GA client-side
Server-side we just want to include the analytics code in case; once
the page is rendered on the client, we first check we are going to
track (i.e SITE_CONFIG['google_analytics_script_id'] is not empty -
see `config/site.yml`) and then mae sure the analytics category is
authorised: by default this is false so we won't ever track a user
without their explicit consent.

Arguably, a page reload is necessary to start the tracking.