fix: let superadmins invite team members to their own team
A super-admin may be part of other organisations and try to invite
team members on their own team, going through the normal org flow (as
opposed to the omniscient superadmin one). In that scenario, we can't
assume an organisation ID was passed and need to fallback on the
current organisation.