Fix the submission of forms created with `form_with`.
These forms got broken, ever since jQuery UI library was added (for
password strength indicator). This is because Rails forms created with
`form_with` are remote (i.e. AJAX-based) by default. Due to this, when
jQuery UI was introduced to the codebase, it also brought with it AJAX
event handlers for such forms, which hijack form submissions in order to
facilitate JS processing.

In this commit, we explicitly marks forms, which were supposed to be
submitted locally (i.e. without AJAX), as local, so that their
submission isn't hijacked.