Get rid of `status` check in password validation API response handler (in JS).
This is because, for some weird reason, `status` field had a value of empty string, on production. Due to
this flakiness, it's best not to rely on it, and as our existing checks for `score` being a number should suffice.
It's also a success event handler anyway, so we should be good in theory.