Relax the password strength requirement to a more sensible level.
We used to have a score of `4`, which was very high and not really necessary. This is because
we already have account locking, password hashing and 2FA in place. With these measures in place,
it's unrealistic for someone with reasonable skills to be able to hack passwords with scores of
even `2` and above easily.

The [zxcvbn's](https://github.com/dropbox/zxcvbn) docs provide a rough guidelines on how easy it
is to guess such passwords. This commit is an attempt to err on the side of caution (by opting
for a score of `2` instead of `1`), because some passwords with a score of `1` were found to be
quite short, and getting a score of `2` wasn't found to be onerous.