Implement correct fix for blobs on move to ActiveStorage 6.1.
The previously applied fix of renaming the BlobsController, and specifying
it in the routes file was wrong. It worked, but was causing other issues.

On further investigation, the reason the BlobsController was no longer
overriding automatically was due to that controller being renamed to
Blobs::RedirectController.

On overriding the correct class, normal behaviour is resumed.