Unpin Sprockets gem and upgrade to version 4.
We had this pinned to version 3.7.2 because of a segmentation
fault that occurred when upgrading to 4.0+

After investigation it seems that Sprockets now uses the manifest.js
file to compile assets instead of the assets.rb file.

See: https://github.com/sass/sassc-rails/issues/122 and
https://github.com/rails/sprockets/blob/master/UPGRADING.md#manifestjs

This commit upgrades Sprockets and now uses the manifest.js file
to precompile assets which ensures the segmentation fault no
longer occurs.