Fix search in the locations view.
The search box to find a location on postcode was broken when the
paging feature was introduced. This feature used to depend on Javascript
and would only work on locations that were displayed on the current page.

This commit now enables search without javascript by loading a new page
from the server