Add more mail related gems to the Gemfile
net-smtp was added when upgrading to Ruby 3.1.1, as the tests failed
without it. This wasn't sufficient though to get the app running in
the production Rails environment though, since eager loading is enable
there, and the mail gem also tries to require net/imap and net/pop.

I think there's something wrong with how ActionMailer is being handled
here, as I'm not aware that the app actually uses the mail gem, let
alone SMTP, POP or IMAP, I think GOV.UK Notify is used for all the
sending of emails.

Anyway, as a quick fix, just include more gems in the Gemfile.