Address some of the inconsistency with managing super admin users
Move towards managing super admins in the users table, rather than via
organisations.

This commit includes a migration to sync the users with is_super_admin
= true with the members of the super admin organisation.

On top of the data changes, this commit tweaks the interface to not
have the super admin pages depend on the super admin organisation.