Remove difference between super admin and regular admin wrt redirection logic.
Remove the restriction of requiring and organisation from as many pages as
possible: home page, log pages, all admin pages and some actions regarding
updating team membership.

The tests are updated so that super admins are no longer part of an organisation,
reflecting the fact this isn no longer necessary.