Simplify the logic to choose which link in the side nav to highlight
This logic worked by comparing the current path with
a group of strings. It is very difficult to show that this
logic ends up being correct. Adding a new route also runs the
risk that the logic breaks and confusing behaviour may result.

The new logic simply remembers the last side nav clicked and
stores it in the session.