Add conditionals to the radiobuttons of the log selection screen
Selecting which log to show displayed a set of radio buttons, which took
the user to a new screen which allowed the user input further selection
criteria.

This has been replaced by one screen with conditionals to the radio button
actions, so that the admin user can search directly from the log page rather
than progressing to a 'holding' search page.