Add former flexibility in JS execution defaulted-out by new CSP
The addition of a Content Security Policy prevents, by default,
the execution of inline scripts from other domains and the
evalution of strings as JS. The intention of the CSP (for now)
was simply to restrict which domains scripts could be downloaded
and executed from.

This commit re-introduces the ability to do both.