Fix Location and Organisation factories
Traits in the Location and Organisation factories were written in such a way
that they would require a reload for their children to be loaded in memory. This
commit ensures that is no longer necessary.

The Organisation name is now taken directly from the register. This also requires
the order of the stubbing out of the register and the loading of the factories to
be reversed