Unsure the /update_ips feature test passes reliably.
Ip addresses can't be private for the form to add IP addresses to
be valid. Once in a while, the method Faker:Internet#public_ip_v4_address
would produce a private IP address causing the form to be invalid and
the test to fail.

This change ensure that only public IP addresses are chosen for the
test