Ensure creating a user is wrapped in a transaction in rspec test
Upgrading Rspec to version 6 caused this test to fail. The reason is that the 'let' block
where the admin_user is created is called when the 'name' and 'email' let block is loaded.
This causes the user to be created outside the transaction that wraps each test and so,
after this test is run, the user is not cleaned up and deleted from the database. Further
tests will fail because uniqueness constraints will be violated