Add migrations to add and initialise sent_first_ip_survey flag
A survey has to be sent whenever a new admin adds an IP.
This has to done only once so a flag is set to determine
if a survey has already been sent.

The flag is initialised to 'true' if they are a member of
at least one organisation with one allocated IP address.