Add Link to download admin user data
A link was added to download a CSV file containing data
of all the admin users in the system.

It records the email, the name, current sign in date. last sign in date
a, the number of sign ins and the organisation name

If there are multiple organisations for a user, there will
be duplicate rows, with only the organisation name being different